% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIP.R
\name{getIP}
\alias{getIP}
\title{An apache access_log parser}
\usage{
getIP(logfile = "access_log")
}
\arguments{
\item{logfile}{Defaults to "access_log".}
}
\description{
This function reads an apache access_log file and returns
the IP addresses and the reformated access date & time.
It is designed to be used with a Geolocator package to determine that
country of origin of each website access and the time.
}
\examples{
getIP("logdelta")
getIP("maphttp/logdelta")

}
\keyword{access_log}
\keyword{apache}
\keyword{httpd}
