% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdelta.R
\name{logdelta}
\alias{logdelta}
\title{logdelta will return the difference of two files}
\usage{
logdelta(logfile1 = "old_access_log", logfile2 = "access_log")
}
\arguments{
\item{logfile1}{This is the old access_log from apache webserver}

\item{logfile2}{This is the current logfile from apache webserver}
}
\description{
A file name "logdelta" that can be passed to a function is returned.
Data is saved in "logdelta" which are the last n lines of logfile2.
}
\examples{
logdelta("old_access_log", "access_log")
logdelta("maphttp/old_access_log", "maphttp/access_log")
}
\keyword{access_log}
\keyword{apache}
\keyword{httpd}
